Dodanie daty zwrotu ksiazki, naliczanie i zwrot kaucji

+++ Biblioteka/Biblioteka/KontoCzytelnika.h
@@ -7,10 +7,9 @@ private:
 	string pesel;
 	string imie;
 	string nazwisko;
-	double kaucja = 0.0;
+	float kaucja = 0.0;
 	int iloscKsiazek = 0;
 public:
-	//KontoCzytelnika();
 	KontoCzytelnika(const string& pesel, const string& imie, const string& nazwisko, const float& kaucja, const int& iloscKsiazek);
 	KontoCzytelnika(const string& pesel, const string& imie, const string& nazwisko);
 	KontoCzytelnika(const string& pesel);
